♻️ ボタンをコンポーネント化
